cart_service/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
dj-database-url==2.1.0
psycopg2-binary==2.9.9
pymysql==1.1.0
requests==2.31.0
django-cors-headers==4.3.1
